Added Russian translation
Used AI and personal skills to translate and edit. This should be accurate enough. There may be refinements in the future, but it's not certain.

Since Russian and English have quite a few differences, I can't claim for sure that everything is translated as it should be. I checked with third-party translations and it turned out this way.